snmp.conf move to libsnmp-base
If a system has libsnmp installed but not snmp cli installed, then the
snmp.conf file is missing and anything using the library may get a flood
of error messages.

The configuration file is not a /usr/bin/snmp configuration file, but a
configuration file for the library as there may be things that use the
library outside the snmp package.

References:
 https://bugs.debian.org/914657