Add Vyatta variant of the libnetsnmpagent library
This change causes an additional libnetsnmpagent-vyatta library to
be built, based entirely on libnetsnmpagent. This allows us to (eg.)
introduce breaking API changes into the new library without affecting
the majority of clients which aren't aware of such changes, and don't
need to be.

Currently libnetsnmpagent and libnetsnmpagent-vyatta are indentical.
Changes can be introduced based on whether VYATTA_SNMP is defined.

VRVDR-51348