Add context to findVar only in Vyatta lib
The addition of an extra argument to findVar can cause problems when
attempting to build pristine upstream components against the version
of libsnmp provided by this repository.

To avoid this we only make this change in libnetsnmpagent-vyatta
which is (currently) identical to libnetsnmpagent except for this
additional context argument to findVar.

For clients which are aware of the context argument they must link
against libnetsnmpagent-vyatta and define VYATTA_SNMP.

Clients which are not aware will require no changes and can continue
to link against libnetsnmpagent.

This change has the advantage of not requiring updates to all the
agents which reside in this repository, since they do not need the
context, therefore reducing the patch size.

VRVDR-51348