Suppress further unwanted SNMP log messages.
When a v2 trap is generated containing a Counter64 value, warnings
are produced in the journal when an attempt is made to convert to v1.
This is not a problem condition since v1 traps do not support this type.
Suppress the noisy warnings in this case.

VRVDR-51616

(cherry picked from commit b1ef192417d819b3edfebd266c0bf75e7b455639)