SIAD fails to send coldStart trap on system reboot
If the trap receiver is unreachable when snmpd
starts up, the coldstart trap would be lost. To
verify that the trap was sent by snmp daemon,
a flag, is_coldstart, is added to netsnmp_pdu
structure which is set for coldStart trap PDU.
This flag is checked when trap is successfully
sent to log a message.

VRVDR-47729